Added test with correctly orientated nacelle guides
The normal of a surface patch is determined by a right handed coordinate
system attached to the (u,v) = (0,0) point of the surface patch. Here
the tangent of the profile defines the x-direction and the tangent of
the guide defines the y-direction. The normal is the z direction.

As a result, the resulting shell is correctly orientated if

- The direction of rotation of the wing profiles is the y-direction, and
  the wing guides run from the fuselage to the wing tip (in y-direction)

- The direction of rotation of the fuselage profiles is negative x-direction, and
  the fuselage guides run in x-direction (or vice versa)

- The direction of rotation of the first (upper) nacelle profiles is in
  y-direction and the direction of rotation of the nacelle guides is in negative
  x-direction.

Closes: #132