Fixed toolbar glitches on mac os
On mac, the qt toolbar has glitches under qt 5,
if one or more native windows are used. The OpenGL
windows uses a native window and is provokes this
Qt bug. A workaround is, to use the unified toolbar
and set the Qt::AA_DontCreateNativeWidgetSiblings
attribute for the app. More informations to this
bug can be seen here:
 - https://bugreports.qt.io/browse/QTBUG-41257
 - https://bugreports.qt.io/browse/QTBUG-39210

fixed #143