Fixed VTK metadata with extended flaps
When the flaps were extended, the mesh points on the flap
could not be connected to the wing geometry and thus did
not contain valid segment, eta, xsi data. Now, we explicitly
store temporary transformations as face traits. Thus, we can
back-transform the points and map them onto the wing.