Workaround the OCCT fuse problems by explicit modeling of TE
The wing-fuselage fusing fails on some models. The reason is an irregular
parametrization of the trailing edge, caused by the MakePatches algorithm.
In this workaround, only the aerodynamic surface of the wing is build
using MakePatches. The trailing edge is modeled using a rules surface loft
from the upper leading edge to the lower leading edge.

All unit tests pass. Also all confidential tests seem to pass.

Addresses #170