Fixed bug in wing span computation
As we have no semantics in CPACS, what kind of wing (HTP, VTP ...)
we have, it is hard to identify the direction of evaluation for
the span computation. Until now, we just use the direction of the
wing whose bounding box is largest. That could produce errors, if
the depth of the wing is larger than its span. With the current
implementation, we ignore the depth by just looking at the leading
edge. We now choose the "major" direction of the leading edge.

Fixes #185