Fixed issue in tiglWingGetSegmentEtaXsi
In cases, when the airfoil is completely above the chord
surface, a point on the cord surface was marked outside
of the wing segment. This several transformations between
component segment and segment could not performed correctly.
In this patch, I am checking also, if the point lies on the
cord surface.