Fixed a bug in segment volume computation
When opening a config via TiGL Viewer, the fuselage
segment volume could be computed correctly. When using
tigl directly it was 0. The reason was, that 0 was the
default value which was computed only, when the loft was
generated once. Now, the loft is generated, before returning
the volume.
I also added the function wingGetSegmentVolume to TiGL Viewer.

Fixes #230