bugfix in CTiglMemoryPool
This commit fixes an allocation issue in CTiglMemoryPool: The class took a char*, pushed it to the back of a std::vector<std::string> and returned a pointer to the the c_str() of the last element of the vector to keep the memory valid. It was possible to invalidate the returned pointer, if further char*s get added to the MemoryPool such that the std::vector reallocates. This bug is fixed by using a std::list instead of a std::vector.