Added API func for reparam.+support closed curves
The method reparametrizeBSplineContinuouslyApprox is committed which
reparametrizes B-spline curves as far as possible without changing
its geometry in a continuous way.
Closed curves are know supported for createGordonSurfaceGeneral.
An working algorithm for sorting the input curves of a B-spline
curve network and their intersection parameters is used in the
Gordon Surface methods, i.e., an unsorted curve network is supported
now as well.