Updated to new version of CPACSGen, improved choice validation (#293)
* improved choice validation
* updated submodule to point to newest CPACSGen
* deleted some table information that is now compiled into CPACSGen

The new CPACSGen now supports multiple input schema files and multiple subdirectories in the input directory. For each subdirectory, an additional namespace is generated and the generated files are placed into a subfolder in the output directory. This mechanism is intended to support additional schema files for writing extensions under the <toolspecific> node in CPACS.
For more information see the README.md at https://github.com/RlanderRISCSW/cpacs_tigl_gen

I also added an improvement on the choice validation code.
We can now handle cases where all elements of one alternative of a choice are all optional (e.g. wingStringerType).
Furthermore the code is now capable of determining the exact alternative of a choice currently represented by the set class members. This may be used in the future to auto detect a choice when reading, generate code for querying the current alternative represented by the class members and switching between alternatives.