Bugfix of tiglFuselageGetPointAngle and better performance (#295)
The function GetPointAngle now takes the cross section center as starting point for the angle,
rather than the origin of the yz-plane at the current eta. 
Also, the performance of the function was improved by roughly a factor of 30. This commit
reintegrates the function getCenterOfMass from the cpacs_3 branch which avoids a planar
section, but uses a iso-parametric eta curve. The run time per call is reduced from about
39 ms to 1.2 ms.

Fixes #264