Fix python linking on macOS (#428)
* Try to fix linking python library

On macOS with anaconda, I got the error PyThreadState_Get: no current thread.
The reason is, that Anacondas Python links libpython statically. Hence,
the tigl extensions must not link a different python library.
To deal with it in a generic fashion, I ask the Python interpreter for
the correct linking args.