Allow to use geom_curves for surface creation
When using the python api I realized, that I could crash
the surface builders, when not providing b-spline curves,
but e.g. trimmed curves.

This commit makes the use of the surface factories a bit
more generic. They can now be used for arbitrary curves.
The b-spline conversion is performed i a first step.

closes #485