Fix computation of mass center for infinitely small shapes
A zero section size, which can occur at the tip or the tail of the aircraft
will result in an infinitely small fuselage cross section wire.
The OpenCASCADE function to compute the center of mass will return the point 0,0,0
which in the end will lead to a roughly double fuselage center line length.

This fix will work around the OpenCASCADE issue by taking account the shape's
bounding box center for very small shapes.

Fixes #551