Fixed invalid cpacs after write
Before, the guide curve fromCircumeferenceValue was set
by TiGL after connecting the guide curves. If a curve had set
fromGuideCurveUID, both values were written to XML, which
is not allowed.

Now, I added the method CCPACSGuideCurve::GetFromRelativeCircumference
such that an external code can ask for the circumference value, even
if fromGuideCurveUID is set. This was suggested by Roland.
There are many benefits by this approach:
 - More functional approach with less side effects
 - Fixed invalid xml
 - Simpler use of the guide curve. Simplifies calling code
 - Better code design that adheres better
    to the single responsibilty principle

    Closes #665