Respect PointBehaviour for tiglWingGetSegmentEtaXsi
I had to change the class structure / method signature slightly
to allow to pass, how GetEtaXsi is evaluated.

Unfortunately, GetEtaXsi on the paramter surfaces is almost
a factor of thousand slower, which could be a problem for the
VTK export with metadata. Therefore I decided to use the old
eta / xsi interpretation for the exports for now.

To validate, I changed the consistency test, added a performance
test and added a function to tiglviewer to play around with
the function.

closes #808