Reactivate 'NiceKnots' in BuildWires()
When there are parameters defined in the CPACS profile, the reparameterization in CTiglInterpolatePointsWithKinks() does not get along with the following reparametrizeBSplineNiceKnots() which was implemented for performance reasons. In this case, the nice knots change the geometry in a too drastic way (cf. unittest BSplineInterpolation.reparameterizePiecewiseLinear). To still account for these performance advantages in most cases, the function is reactivated. But only if there are no parameters defined in the CPACS profile.