Throw error when sections are added in segments containing gc
The CPACS creator currently allows adding sections in segments (or at their boundaries) when they contain guide curves. However this leads to crashes or undefined behaviour. As it might be hard to discuss reasonable behaviour or predict what the user wants in complex examples, we do not allow this right now.
(At least partly) fixes issue #626