Test whether file was changed before saving
Logic: When CPACSConfiguration is changed by user, undoCommandRequired() signal is emitted. This in turn implicitely activates TIGLViewerWindow::updateScene(). Here, a bool variable is set which stores information whether the file is edited after last save or not.
The SaveAs() command is not influenced by this variable. It should always be possible. Before executing saveFile(), it is checked by means of the new variable whether there were changes.
If the current CPACSConfiguration is closed, the new Do-you-want-to-save-dialog is only activated when there were changes. In case of no changes, there is no data loss when not asked before.