Restructure the template files to avoid wrong indenting (#1120)
Currently, in some cases when files are built up from templates, the final CPACS file is not correctly indented.
As a consequence, the resulting file can be hardly human-readable. This results from xml-behaviour. Here, whitespace CAN be relevant for the interpreter.
That is why, xml does not simply create them (which would be needed for correct indenting). To get xml to do so, the template files are written in just one line.
When the user stores the final edited result, xml will then indent it correctly since there were no whitespaces before.