Fix #1108; Choose width as target value instead of area
When adding a new section within a wing, the area was linearly interpolated as half of the sum of the start und end element. This area was used as a target value to define the new wing. Since the relation between area and wing width is not linear, the width of the new element does not fit will into the existing wing. As a result, a kink is created on the wing's leading and trailing edge. The scaling/translation appears wrong or at least not meaningful. When changing the target value to the wing width directly, the leading and trailing edge point is on the yet existing respective edge. That means that the width is interpolated linearly and not the area anymore. Hence, no kink appears anymore and the wing's shape does not change. As the used SetWidth() function does also slightly move the center, one test had to be adapted.