Change interpolation of new added fuselage element according to changes for wing
Although the result will not look as neat as for wings, we change the interpolation for consistency. The reason for different results is that the fuselage lofting is not linear in general. In fact, the lofting algorithm uses B-Splines. However, the changed interpolation for the new element assumes linear lofts when it should keep the geometry (based on findings for wings in #1108).