Change default value in transparency user widget to 0
The current default does not seem reasonable. When the user activates the widget to change the value, the hard-wired value 30 was pre-seleted. That way, when the user leaves the widget (without actively changing this value) the opacity nevertheless is set to 30, since this value is selected. Here, a value of 0 seems more sensible. However, in future, this value should be stored within the scence graph. That way, it can be read out each time the widget is called and behaves like expected.