Extend definition of approximationSettings node
Additionally added options: User can define the approximation error method, currently: root mean square error or maximum error. Furthermore, the user can pass an index list within the CPACS configuration to define which points should still be interpolated. The other points, however, are approximated. The unittests needed to be changed to test for this behaviour. Also, a mistake was fixed, since the was a confusion regarding 0-based and 1-based indexing within th logic of interpolation indices.